API-6841 - Move to separate module
