+++ app/uk/gov/hmrc/apiplatformmicroservice/commands/applications/config/ApplicationCommandConnectorConfigurationProviders.scala
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.apiplatformmicroservice.thirdpartyapplication.config
+package uk.gov.hmrc.apiplatformmicroservice.commands.applications.config
 
 import com.google.inject.{Inject, Provider, Singleton}
+import uk.gov.hmrc.apiplatformmicroservice.commands.applications.connectors._
 
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
-import uk.gov.hmrc.apiplatformmicroservice.thirdpartyapplication.connectors.{SubordinateApplicationCommandConnector, PrincipalApplicationCommandConnector}
-
+import uk.gov.hmrc.apiplatformmicroservice.thirdpartyapplication.config.ConfigProviderHelper
 
 @Singleton
 class PrincipalApplicationCommandConnectorConfigProvider @Inject() (override val sc: ServicesConfig)
