+++ app/uk/gov/hmrc/apiplatformmicroservice/common/utils/EitherTHelper.scala
@@ -24,7 +24,7 @@ import cats.instances.future.catsStdInstancesForFuture
 trait EitherTHelper[E] {
   implicit val ec: ExecutionContext
 
-  def pure[A](in: A)                                                            = EitherT.pure[Future, String](in)
+  def pure[A](in: A)                                                            = EitherT.pure[Future, E](in)
   def liftF[A](in: Future[A]): EitherT[Future, E, A]                            = EitherT.liftF(in)
   def fromOption[A](in: Option[A], error: => E): EitherT[Future, E, A]          = EitherT.fromOption(in, error)
   def fromOptionF[A](in: Future[Option[A]], error: => E): EitherT[Future, E, A] = EitherT.fromOptionF(in, error)
