+++ app/uk/gov/hmrc/apiplatformmicroservice/thirdpartyapplication/config/ConfigurationProviders.scala
@@ -37,8 +37,5 @@ class ConfigurationModule extends AbstractModule {
     bind(classOf[SubscriptionFieldsConnector]).annotatedWith(named("principal")).to(classOf[PrincipalSubscriptionFieldsConnector])
 
     bind(classOf[ThirdPartyDeveloperConnector.Config]).toProvider(classOf[ThirdPartyDeveloperConnectorConfigurationProvider])
-
-    bind(classOf[PrincipalApplicationCommandConnector.Config]).toProvider(classOf[PrincipalApplicationCommandConnectorConfigProvider])
-    bind(classOf[SubordinateApplicationCommandConnector.Config]).toProvider(classOf[SubordinateApplicationCommandConnectorConfigProvider])
   }
 }
