--- app/uk/gov/hmrc/apiplatformmicroservice/thirdpartyapplication/services/PassThruDispatcher.scala
@@ -1,40 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.apiplatformmicroservice.thirdpartyapplication.services
-
-import uk.gov.hmrc.apiplatformmicroservice.thirdpartyapplication.domain.models.applications.Application
-import javax.inject.{Inject, Singleton}
-
-import uk.gov.hmrc.apiplatform.modules.commands.applications.domain.models._
-import uk.gov.hmrc.apiplatformmicroservice.thirdpartyapplication.connectors.EnvironmentAwareApplicationCommandConnector
-import cats.data.NonEmptyChain
-import scala.concurrent.ExecutionContext
-import uk.gov.hmrc.apiplatformmicroservice.common.utils.EitherTHelper
-import uk.gov.hmrc.http.HeaderCarrier
-
-@Singleton
-class PassThruDispatcher @Inject() (
-    val connector: EnvironmentAwareApplicationCommandConnector
-  )(implicit val ec: ExecutionContext
-  ) {
-
-  val E = EitherTHelper.make[NonEmptyChain[CommandFailure]]
-
-  def dispatch(app: Application, dispatchRequest: DispatchRequest)(implicit hc: HeaderCarrier): ApplicationCommandHandlerTypes.Result = {
-    connector(app.deployedTo).dispatch(app.id, dispatchRequest)
-  }
-}
