+++ conf/application.conf
@@ -51,6 +51,7 @@ play.modules.enabled += "uk.gov.hmrc.apiplatformmicroservice.xmlapis.config.Conf
 play.modules.enabled += "uk.gov.hmrc.apiplatformmicroservice.metrics.Module"
 play.modules.enabled += "uk.gov.hmrc.apiplatformmicroservice.pushpullnotifications.config.ConfigurationModule"
 play.modules.enabled += "uk.gov.hmrc.apiplatformmicroservice.thirdpartyapplication.config.ClockModule"
+play.modules.enabled += "uk.gov.hmrc.apiplatformmicroservice.thirdpartyapplication.commands.applications.config.ConfigurationModule"
 
 # Session Timeout
 # ~~~~
